+++ IHP/ViewPrelude.hs
@@ -33,7 +33,7 @@ module IHP.ViewPrelude (
     module IHP.LoginSupport.Helper.View,
     module IHP.PageHead.ViewFunctions,
     module IHP.Assets.ViewFunctions,
-    module IHP.Breadcrumbs.ViewFunctions,
+    module IHP.Breadcrumb.ViewFunctions,
     module IHP.Pagination.ViewFunctions,
 ) where
 
@@ -68,5 +68,5 @@ import IHP.LoginSupport.Helper.View
 import IHP.PageHead.ViewFunctions
 
 import IHP.Assets.ViewFunctions
-import IHP.Breadcrumbs.ViewFunctions
+import IHP.Breadcrumb.ViewFunctions
 import IHP.Pagination.ViewFunctions
